+++ tests/Fake/ConfigurableTraitUnconfigured.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Ork
+ * Ork Core
  *
  * @package   Ork\Core\Tests
  * @copyright 2015-2021 Alex Howansky (https://github.com/AlexHowansky)