+++ tests/String/Base32Test.php
@@ -1,9 +1,9 @@
 <?php
 
 /**
- * Ork
+ * Ork Core
  *
- * @package   Ork\Core
+ * @package   Ork\Core\Tests
  * @copyright 2015-2021 Alex Howansky (https://github.com/AlexHowansky)
  * @license   https://github.com/AlexHowansky/ork-core/blob/master/LICENSE MIT License
  * @link      https://github.com/AlexHowansky/ork-core